elimination_task/requirements.txt
numpy~=1.21.5
Pillow~=8.4.0
scipy~=1.7.3